# Add phrase_from_stream/2 to library(pio)