I wanted to learn how to use global variables in Prolog (`bb_get/2` and `bb_put/2` specifically), and thought this was a good application. It was very, _very_ hard but I think I did it.

In my opinion `phrase_from_stream/2` is the most useful predicate that wasn't yet implemented. I already stumbled into many situations where it would have been incredibly useful, mainly dealing with sockets.

In this implementation it still uses the old `stream_to_lazy_list/2` for repositionable streams, as I bet it's way more efficient than whatever shenanigans I'm doing in the other case, especially for large files. For non-repositionable streams it stores the data from the string in a buffer to enable backtracking, as suggested in #1146.

I'm not sure if this can handle arbitrary binary data, as it uses `partial_string/3` and `partial_string_tail/2`, but the old `stream_to_lazy_list/2` also used `partial_string/3`, so at least I'm not regressing here.

The buffer is a partial string, and I'm not sure what this means for the performance of this for long streams. I'm currently using a custom `string_get_n_chars/4` implementation that is recursive and therefore _does not_ acess the buffer with random access. I know that Scryer uses a very efficient internal representation for partial lists, so maybe there is some way to get a substring more efficiently than this if this ever presents itself as a important performance problem in the future.

Closes #1146.