@triska 2023-09-22
Incredible, thank you a lot for working on this!

There is a very efficient internal predicate for skipping elements in lists, called **`'$skip_max_list'/4`**, see #17 and #1023. It seems that could help for your use case, to skip a number of elements directly in memory without any Prolog-based overhead.