@bakaq 2023-09-22
> There is a very efficient internal predicate for skipping elements in lists, called **`'$skip_max_list'/4`**

That's definitively a predicate that is good to know. I worry that maybe because it's an internal predicate that may hinder the portability of this library, but it already uses freeze and now global variables, which aren't ISO as far as I know, and it isn't that difficult to create an ISO compliant alternative (that's not as efficient, like I did) if there is no equivalent predicate in another Prolog system.