@triska 2023-09-22
Yes, that, and also: All other Prolog systems will also benefit from adding this internal predicate! It is in fact one of the most important low-level building blocks for writing correct and efficient library code in Prolog.