@bakaq 2023-09-22
Are there any plans to put something like it in the standard, or it's too much of an implementation detail for that?