@UWN 2023-09-22
> Are there any plans to put something like it in the standard, or is it too much of an implementation detail for that?

Effectively, skip_max_list is a very low level predicate that reveals also some implementation details that are otherwise not accessible (and where it does not make sense to provide them). Also, these differences only show for implementations that opt for (rational) trees for NSTO unifications. And there are many tiny details to sort out that are different in various implementations.

```
?-  maplist(=(a), As),Rep=[_|_],append(Pre, Rep, As), append(Rep, Loop,Loop), append(Pre, Loop, L).
´´´
There is only one value for `L`, but skip_max_list would produce all the different representations.