@bakaq 2023-09-23
With `'$skip_max_list/4'` this now is much faster. I benchmarked reading a [big json file](https://www.reddit.com/r/linux.json) (about 160KB) with `json_chars//1` from `library(serialization/json)`. I measured the time to get the result, and the time to get the extra `false` from the choice point `json_chars//1` leaves, terminating the query.

| Method  | Time for result | Time to `false` | Total time to termination |
| --- | --- | --- | --- |
| Old method for repositionable streams  | 44.9s   | 60.7s     | 105.6s |
| Original version of this PR                            | 268.5s | 378.7s   | 647.2s |
| New version with `'$skip_max_chars'/4` | 63.7s    | 87.3s     | 151.0s |

The total time to termination is what one would actually see in a program unless they used something like `once/1` or a cut to avoid the choice point. The performance improved about 4.2 times, and it's now only about 1.4 times slower that the version for repositionable streams (and for smaller files it gets even closer!). I believe that it can actually be _faster_ than the repositionable version (because I think repositioning a stream needs a syscall), but then `string_get_n_chars/4` would need to be optimized. It currently is a Prolog implemented recursive predicate, and it needs to recurse 4096 times almost every single call in the current version. Any ideas on how to improve this?