@pmoura 2023-09-23
> (I know this isn't a fair comparison, but Python's json library parses this same file in about 868 microseconds)

For a more fair comparison, you can compare with Logtalk's portable [json](https://logtalk.org/manuals/libraries/json.html) library:

```text
$ scryerlgt
...
?- {json(loader)}.
...
?- use_module(library(time)).
   true.
?- time(json(chars)::parse(file('linux.json'), T)).
   % CPU time: 0.637s
   T = {chars("kind")-chars("Listing"),chars("data")-{chars("after")-chars("t3_16pckfq"),chars("dist")-26,chars("modhash")-chars([]),chars("geo_filte ...")- @null,chars("children ...")-[{chars("kind")-chars("t3"),chars("data")-{chars("ap ...")- @null,chars("s ...")-chars("l ..."),chars("selftext")-chars("If you\'re working on an interesting Linux based project, here\'s the place to post it. Share what you\'ve done."),chars(...)-chars(...),chars(...)- ...,...,...}},{chars("kind ...")-chars("t3"),chars("data ...")-{chars("a ...")- @null,chars("subreddit")-chars("linux"),chars(...)-chars(...),chars(...)-chars(...),...,...}},{chars("kin ...")-chars("t3"),chars("dat ...")-{chars("approved_at_utc")- @null,chars(...)-chars(...),chars(...)-chars(...),...,...}},{chars("ki ...")-chars("t3 ..."),chars("da ...")-{chars(...)- @ ...,chars(...)-chars(...),...,...}},{chars("k ...")-chars("t ..."),chars("d ...")-{chars(...)- ...,...,...}},{chars("kind")-chars("t3"),chars("data")-{...,...}},{chars(...)-chars(...),chars(...)-{...}},{chars(...)-chars(...),chars(...)-{...}},{...,...},{...}|...],chars("before")- @null}}.
```

Above time with the latest Scryer Prolog git version (1c8cd85f6c0c429c632ae62f1cbedee17b998554) running on an Intel iMac.