@bakaq 2023-09-23
> For a more fair comparison, you can compare with Logtalk's portable [json](https://logtalk.org/manuals/libraries/json.html) library

I don't think that's fair either. Scryer's `library(serialization/json)` uses DCGs, and there's only one predicate that works in both directions, while Logtalk's json library seems to have different predicates for parsing and generating. Having different predicates for that enables some quite obvious optimizations and even a naive implementation will still probably be faster than Scryer's current json library (#1566), but having a DCG non-terminal that describes JSON enables some pretty interesting stuff. It's kind of comparing apples to oranges. DCGs are more flexible for reasoning, but modal predicates are way more performant.

Logtalk's json library wouldn't need nor benefit from `phrase_from_stream/2`, because it's not a DCG. I used `json_chars//2` for benchmarking because it's probably the most complex DCG in the Scryer Prolog libraries. Also, there's certainly a lot of optimization on the table there.