@pmoura 2023-09-23
The Logtalk `json` library uses separate DCGs for parsing and generating. The top non-terminals are not currently part of the **public** interface. They could easily be (it's just a question of adding a scope directive for the non-terminals) but there was never a need for it.