@triska 2023-09-24
One other question I have, also related to the alternative execution strategy I mentioned, using attributed variables: Does `phrase_from_stream/2` leave any residual terms in memory, so that more and more memory is consumed? One simple test case would be to call `phrase_from_stream/2` repeatedly on a suitable stream, for example as:

    repeat, setup_stream(Stream), phrase_from_stream(seq(_), Stream), false.

and watch the memory consumption of the Prolog process with the program `htop`.

It may be possible to add suitable cleanup to the `setup_call_cleanup/3` call which is already used in the code. The use of attributed variables may make it easier to clean up, because the attributes are automatically reset on backtracking, whereas global&nbsp;variables need to be reset "manually".