@bakaq 2023-09-24
> The use of attributed variables may make it easier to clean up, because the attributes are automatically reset on backtracking, whereas global variables need to be reset "manually".

The problem with attributed variables for this is that I think it would be much harder to implement the buffer, precisely because they are backtrackable. At least in the way I implemented, the buffer not being backtracked is essential, so that we don't lose data which was already read from the stream, as it cannot be read again. I'm not sure something like this is possible with attributed variables. BufferPos on the other hand probably could use attributed variables very easily.